refactor(layout): move template machinery from paginate to page
